refactor(ui): move ModelRefLabel into its own file
