+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/sections/ComparisonDefinitionSection/ComparisonDefinitionSection.tsx
@@ -13,11 +13,7 @@ import {
 import {useEvaluationsFilter} from '../../../CallsPage/evaluationsFilter';
 import {Id} from '../../../common/Id';
 import {opNiceName} from '../../../common/opNiceName';
-import {useWFHooks} from '../../../wfReactInterface/context';
-import {
-  CallSchema,
-  ObjectVersionKey,
-} from '../../../wfReactInterface/wfDataModelHooksInterface';
+import {CallSchema} from '../../../wfReactInterface/wfDataModelHooksInterface';
 import {useCompareEvaluationsState} from '../../compareEvaluationsContext';
 import {
   EvaluationComparisonState,
@@ -29,6 +25,7 @@ import {HorizontalBox} from '../../Layout';
 import {ItemDef} from '../DraggableSection/DraggableItem';
 import {DraggableSection} from '../DraggableSection/DraggableSection';
 import {VerticalBar} from './EvaluationDefinition';
+import {ModelRefLabel} from './ModelRefLabel';
 
 export const ComparisonDefinitionSection: React.FC<{
   state: EvaluationComparisonState;
@@ -87,39 +84,6 @@ export const ComparisonDefinitionSection: React.FC<{
         </HorizontalBox>
       </div>
     </Tailwind>
-  );
-};
-
-const ModelRefLabel: React.FC<{modelRef: string}> = props => {
-  const {useObjectVersion} = useWFHooks();
-  const objRef = useMemo(
-    () => parseRef(props.modelRef) as WeaveObjectRef,
-    [props.modelRef]
-  );
-  const objVersionKey = useMemo(() => {
-    return {
-      scheme: 'weave',
-      entity: objRef.entityName,
-      project: objRef.projectName,
-      weaveKind: objRef.weaveKind,
-      objectId: objRef.artifactName,
-      versionHash: objRef.artifactVersion,
-      path: '',
-      refExtra: objRef.artifactRefExtra,
-    } as ObjectVersionKey;
-  }, [
-    objRef.artifactName,
-    objRef.artifactRefExtra,
-    objRef.artifactVersion,
-    objRef.entityName,
-    objRef.projectName,
-    objRef.weaveKind,
-  ]);
-  const objectVersion = useObjectVersion({key: objVersionKey});
-  return (
-    <span className="ml-2">
-      {objectVersion.result?.objectId}:v{objectVersion.result?.versionIndex}
-    </span>
   );
 };
